Arrays not getting correctly converted to JsonArray
