+++ src/typechat.program/ProgramInterpreter.cs
@@ -252,32 +252,5 @@ public class ProgramInterpreter
     /// </summary>
     /// <param name="obj">obj to convert</param>
     /// <returns></returns>
-    public static JsonNode ToJsonNode(dynamic obj)
-    {
-        if (obj == null)
-        {
-            return obj;
-        }
-        if (obj is dynamic[] darray)
-        {
-            JsonArray jsonArray = new JsonArray();
-            for (int i = 0; i < darray.Length; ++i)
-            {
-                dynamic value = darray[i];
-                jsonArray.Add(ToJsonNode(value));
-            }
-            return jsonArray;
-        }
-        else if (obj is Array array)
-        {
-            JsonArray jsonArray = new JsonArray();
-            for (int i = 0; i < array.Length; ++i)
-            {
-                dynamic value = array.GetValue(i);
-                jsonArray.Add(ToJsonNode(value));
-            }
-            return jsonArray;
-        }
-        return obj;
-    }
+    public static JsonNode ToJsonNode(dynamic obj) => JsonNodeEx.ToJsonNode(obj);
 }
